requirements.txt
Flask
markdown
GitPython